Search words in wordbook
