+++ pubspec.yaml
@@ -80,6 +80,7 @@ dependencies:
   tray_manager: ^0.5.0
   window_manager: ^0.5.1
   hotkey_manager: ^0.2.3
+  fuzzy: ^0.5.1
 
 dev_dependencies:
   flutter_test:
